enable some useful jekyll plugins
Once I already disabled jekyll-feed, because I wanted
not quite a blog, but rather a personal website,
and a news feed wasn't very useful.

And RSS feeds wasn't seem very useful to me back then.
Nowadays I use news aggregator a lot, and now it is
ovbvious how useful it could be, so I'm re-enabling it.